@@ -5,7 +5,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
+// 	protoc-gen-go v1.36.5
 // 	protoc        (unknown)
 // source: proto/extension/context.proto
 
@@ -16,6 +16,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -28,23 +29,20 @@ const (
 // RouteExtensionContext provides resources introduced by an extension and watched by Envoy Gateway
 // additional context information can be added to this message as more use-cases are discovered
 type PostRouteExtensionContext struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// Resources introduced by the extension that were used as extensionRefs in an HTTPRoute/GRPCRoute
 	ExtensionResources []*ExtensionResource `protobuf:"bytes,1,rep,name=extension_resources,json=extensionResources,proto3" json:"extension_resources,omitempty"`
 	// hostnames are the fully qualified domain names attached to the HTTPRoute
-	Hostnames []string `protobuf:"bytes,2,rep,name=hostnames,proto3" json:"hostnames,omitempty"`
+	Hostnames     []string `protobuf:"bytes,2,rep,name=hostnames,proto3" json:"hostnames,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PostRouteExtensionContext) Reset() {
 	*x = PostRouteExtensionContext{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_context_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_context_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostRouteExtensionContext) String() string {
@@ -55,7 +53,7 @@ func (*PostRouteExtensionContext) ProtoMessage() {}
 
 func (x *PostRouteExtensionContext) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_context_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -88,18 +86,16 @@ func (x *PostRouteExtensionContext) GetHostnames() []string {
 // breaking any clients that use the API
 // additional context information can be added to this message as more use-cases are discovered
 type PostVirtualHostExtensionContext struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
 	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PostVirtualHostExtensionContext) Reset() {
 	*x = PostVirtualHostExtensionContext{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_context_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_context_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostVirtualHostExtensionContext) String() string {
@@ -110,7 +106,7 @@ func (*PostVirtualHostExtensionContext) ProtoMessage() {}
 
 func (x *PostVirtualHostExtensionContext) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_context_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -129,22 +125,19 @@ func (*PostVirtualHostExtensionContext) Descriptor() ([]byte, []int) {
 // breaking any clients that use the API
 // additional context information can be added to this message as more use-cases are discovered
 type PostHTTPListenerExtensionContext struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state protoimpl.MessageState `protogen:"open.v1"`
 	// Resources introduced by the extension that were used as extension server
 	// policies targeting the listener
 	ExtensionResources []*ExtensionResource `protobuf:"bytes,1,rep,name=extension_resources,json=extensionResources,proto3" json:"extension_resources,omitempty"`
+	unknownFields      protoimpl.UnknownFields
+	sizeCache          protoimpl.SizeCache
 }
 
 func (x *PostHTTPListenerExtensionContext) Reset() {
 	*x = PostHTTPListenerExtensionContext{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_context_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_context_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostHTTPListenerExtensionContext) String() string {
@@ -155,7 +148,7 @@ func (*PostHTTPListenerExtensionContext) ProtoMessage() {}
 
 func (x *PostHTTPListenerExtensionContext) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_context_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -181,18 +174,16 @@ func (x *PostHTTPListenerExtensionContext) GetExtensionResources() []*ExtensionR
 // breaking any clients that use the API
 // additional context information can be added to this message as more use-cases are discovered
 type PostTranslateExtensionContext struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
 	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PostTranslateExtensionContext) Reset() {
 	*x = PostTranslateExtensionContext{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_context_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_context_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostTranslateExtensionContext) String() string {
@@ -203,7 +194,7 @@ func (*PostTranslateExtensionContext) ProtoMessage() {}
 
 func (x *PostTranslateExtensionContext) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_context_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -223,20 +214,17 @@ func (*PostTranslateExtensionContext) Descriptor() ([]byte, []int) {
 // can marshal the bytes from this resource back into an unstructured.Unstructured and then
 // perform type checking to obtain the resource.
 type ExtensionResource struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	UnstructuredBytes []byte `protobuf:"bytes,1,opt,name=unstructured_bytes,json=unstructuredBytes,proto3" json:"unstructured_bytes,omitempty"`
+	state             protoimpl.MessageState `protogen:"open.v1"`
+	UnstructuredBytes []byte                 `protobuf:"bytes,1,opt,name=unstructured_bytes,json=unstructuredBytes,proto3" json:"unstructured_bytes,omitempty"`
+	unknownFields     protoimpl.UnknownFields
+	sizeCache         protoimpl.SizeCache
 }
 
 func (x *ExtensionResource) Reset() {
 	*x = ExtensionResource{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_context_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_context_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ExtensionResource) String() string {
@@ -247,7 +235,7 @@ func (*ExtensionResource) ProtoMessage() {}
 
 func (x *ExtensionResource) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_context_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -271,7 +259,7 @@ func (x *ExtensionResource) GetUnstructuredBytes() []byte {
 
 var File_proto_extension_context_proto protoreflect.FileDescriptor
 
-var file_proto_extension_context_proto_rawDesc = []byte{
+var file_proto_extension_context_proto_rawDesc = string([]byte{
 	0x0a, 0x1d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x65, 0x78, 0x74, 0x65, 0x6e, 0x73, 0x69, 0x6f,
 	0x6e, 0x2f, 0x63, 0x6f, 0x6e, 0x74, 0x65, 0x78, 0x74, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
 	0x16, 0x65, 0x6e, 0x76, 0x6f, 0x79, 0x67, 0x61, 0x74, 0x65, 0x77, 0x61, 0x79, 0x2e, 0x65, 0x78,
@@ -304,22 +292,22 @@ var file_proto_extension_context_proto_rawDesc = []byte{
 	0x65, 0x64, 0x42, 0x79, 0x74, 0x65, 0x73, 0x42, 0x11, 0x5a, 0x0f, 0x70, 0x72, 0x6f, 0x74, 0x6f,
 	0x2f, 0x65, 0x78, 0x74, 0x65, 0x6e, 0x73, 0x69, 0x6f, 0x6e, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74,
 	0x6f, 0x33,
-}
+})
 
 var (
 	file_proto_extension_context_proto_rawDescOnce sync.Once
-	file_proto_extension_context_proto_rawDescData = file_proto_extension_context_proto_rawDesc
+	file_proto_extension_context_proto_rawDescData []byte
 )
 
 func file_proto_extension_context_proto_rawDescGZIP() []byte {
 	file_proto_extension_context_proto_rawDescOnce.Do(func() {
-		file_proto_extension_context_proto_rawDescData = protoimpl.X.CompressGZIP(file_proto_extension_context_proto_rawDescData)
+		file_proto_extension_context_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_proto_extension_context_proto_rawDesc), len(file_proto_extension_context_proto_rawDesc)))
 	})
 	return file_proto_extension_context_proto_rawDescData
 }
 
 var file_proto_extension_context_proto_msgTypes = make([]protoimpl.MessageInfo, 5)
-var file_proto_extension_context_proto_goTypes = []interface{}{
+var file_proto_extension_context_proto_goTypes = []any{
 	(*PostRouteExtensionContext)(nil),        // 0: envoygateway.extension.PostRouteExtensionContext
 	(*PostVirtualHostExtensionContext)(nil),  // 1: envoygateway.extension.PostVirtualHostExtensionContext
 	(*PostHTTPListenerExtensionContext)(nil), // 2: envoygateway.extension.PostHTTPListenerExtensionContext
@@ -341,73 +329,11 @@ func file_proto_extension_context_proto_init() {
 	if File_proto_extension_context_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_proto_extension_context_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostRouteExtensionContext); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_context_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostVirtualHostExtensionContext); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_context_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostHTTPListenerExtensionContext); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_context_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostTranslateExtensionContext); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_context_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ExtensionResource); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_proto_extension_context_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_proto_extension_context_proto_rawDesc), len(file_proto_extension_context_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   5,
 			NumExtensions: 0,
@@ -418,7 +344,6 @@ func file_proto_extension_context_proto_init() {
 		MessageInfos:      file_proto_extension_context_proto_msgTypes,
 	}.Build()
 	File_proto_extension_context_proto = out.File
-	file_proto_extension_context_proto_rawDesc = nil
 	file_proto_extension_context_proto_goTypes = nil
 	file_proto_extension_context_proto_depIdxs = nil
 }
