@@ -5,7 +5,7 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.33.0
+// 	protoc-gen-go v1.36.5
 // 	protoc        (unknown)
 // source: proto/extension/service.proto
 
@@ -20,6 +20,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -31,21 +32,18 @@ const (
 
 // PostRouteModifyRequest sends a Route that was generated by Envoy Gateway along with context information to an extension so that the Route can be modified
 type PostRouteModifyRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state            protoimpl.MessageState     `protogen:"open.v1"`
 	Route            *v3.Route                  `protobuf:"bytes,1,opt,name=route,proto3" json:"route,omitempty"`
 	PostRouteContext *PostRouteExtensionContext `protobuf:"bytes,2,opt,name=post_route_context,json=postRouteContext,proto3" json:"post_route_context,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *PostRouteModifyRequest) Reset() {
 	*x = PostRouteModifyRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_service_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_service_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostRouteModifyRequest) String() string {
@@ -56,7 +54,7 @@ func (*PostRouteModifyRequest) ProtoMessage() {}
 
 func (x *PostRouteModifyRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_service_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -88,20 +86,17 @@ func (x *PostRouteModifyRequest) GetPostRouteContext() *PostRouteExtensionContex
 // PostRouteModifyResponse is the expected response from an extension and contains a modified version of the Route that was sent
 // If an extension returns a nil Route then it will not be modified
 type PostRouteModifyResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Route         *v3.Route              `protobuf:"bytes,1,opt,name=route,proto3" json:"route,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Route *v3.Route `protobuf:"bytes,1,opt,name=route,proto3" json:"route,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PostRouteModifyResponse) Reset() {
 	*x = PostRouteModifyResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_service_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_service_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostRouteModifyResponse) String() string {
@@ -112,7 +107,7 @@ func (*PostRouteModifyResponse) ProtoMessage() {}
 
 func (x *PostRouteModifyResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_service_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -136,21 +131,18 @@ func (x *PostRouteModifyResponse) GetRoute() *v3.Route {
 
 // PostVirtualHostModifyRequest sends a VirtualHost that was generated by Envoy Gateway along with context information to an extension so that the VirtualHost can be modified
 type PostVirtualHostModifyRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state                  protoimpl.MessageState           `protogen:"open.v1"`
 	VirtualHost            *v3.VirtualHost                  `protobuf:"bytes,1,opt,name=virtual_host,json=virtualHost,proto3" json:"virtual_host,omitempty"`
 	PostVirtualHostContext *PostVirtualHostExtensionContext `protobuf:"bytes,2,opt,name=post_virtual_host_context,json=postVirtualHostContext,proto3" json:"post_virtual_host_context,omitempty"`
+	unknownFields          protoimpl.UnknownFields
+	sizeCache              protoimpl.SizeCache
 }
 
 func (x *PostVirtualHostModifyRequest) Reset() {
 	*x = PostVirtualHostModifyRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_service_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_service_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostVirtualHostModifyRequest) String() string {
@@ -161,7 +153,7 @@ func (*PostVirtualHostModifyRequest) ProtoMessage() {}
 
 func (x *PostVirtualHostModifyRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_service_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -193,20 +185,17 @@ func (x *PostVirtualHostModifyRequest) GetPostVirtualHostContext() *PostVirtualH
 // PostVirtualHostModifyResponse is the expected response from an extension and contains a modified version of the VirtualHost that was sent
 // If an extension returns a nil Virtual Host then it will not be modified
 type PostVirtualHostModifyResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	VirtualHost   *v3.VirtualHost        `protobuf:"bytes,1,opt,name=virtual_host,json=virtualHost,proto3" json:"virtual_host,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	VirtualHost *v3.VirtualHost `protobuf:"bytes,1,opt,name=virtual_host,json=virtualHost,proto3" json:"virtual_host,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PostVirtualHostModifyResponse) Reset() {
 	*x = PostVirtualHostModifyResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_service_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_service_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostVirtualHostModifyResponse) String() string {
@@ -217,7 +206,7 @@ func (*PostVirtualHostModifyResponse) ProtoMessage() {}
 
 func (x *PostVirtualHostModifyResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_service_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -241,21 +230,18 @@ func (x *PostVirtualHostModifyResponse) GetVirtualHost() *v3.VirtualHost {
 
 // PostVirtualHostModifyRequest sends a Listener that was generated by Envoy Gateway along with context information to an extension so that the Listener can be modified
 type PostHTTPListenerModifyRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state               protoimpl.MessageState            `protogen:"open.v1"`
 	Listener            *v31.Listener                     `protobuf:"bytes,1,opt,name=listener,proto3" json:"listener,omitempty"`
 	PostListenerContext *PostHTTPListenerExtensionContext `protobuf:"bytes,2,opt,name=post_listener_context,json=postListenerContext,proto3" json:"post_listener_context,omitempty"`
+	unknownFields       protoimpl.UnknownFields
+	sizeCache           protoimpl.SizeCache
 }
 
 func (x *PostHTTPListenerModifyRequest) Reset() {
 	*x = PostHTTPListenerModifyRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_service_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_service_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostHTTPListenerModifyRequest) String() string {
@@ -266,7 +252,7 @@ func (*PostHTTPListenerModifyRequest) ProtoMessage() {}
 
 func (x *PostHTTPListenerModifyRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_service_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -298,20 +284,17 @@ func (x *PostHTTPListenerModifyRequest) GetPostListenerContext() *PostHTTPListen
 // PostHTTPListenerModifyResponse is the expected response from an extension and contains a modified version of the Listener that was sent
 // If an extension returns a nil Listener then it will not be modified
 type PostHTTPListenerModifyResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Listener      *v31.Listener          `protobuf:"bytes,1,opt,name=listener,proto3" json:"listener,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Listener *v31.Listener `protobuf:"bytes,1,opt,name=listener,proto3" json:"listener,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PostHTTPListenerModifyResponse) Reset() {
 	*x = PostHTTPListenerModifyResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_service_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_service_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostHTTPListenerModifyResponse) String() string {
@@ -322,7 +305,7 @@ func (*PostHTTPListenerModifyResponse) ProtoMessage() {}
 
 func (x *PostHTTPListenerModifyResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_service_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -347,22 +330,19 @@ func (x *PostHTTPListenerModifyResponse) GetListener() *v31.Listener {
 // PostTranslateModifyRequest currently sends only clusters and secrets to an extension.
 // The extension is free to add/modify/remove the resources it received.
 type PostTranslateModifyRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state                protoimpl.MessageState         `protogen:"open.v1"`
 	PostTranslateContext *PostTranslateExtensionContext `protobuf:"bytes,1,opt,name=post_translate_context,json=postTranslateContext,proto3" json:"post_translate_context,omitempty"`
 	Clusters             []*v32.Cluster                 `protobuf:"bytes,2,rep,name=clusters,proto3" json:"clusters,omitempty"`
 	Secrets              []*v33.Secret                  `protobuf:"bytes,3,rep,name=secrets,proto3" json:"secrets,omitempty"`
+	unknownFields        protoimpl.UnknownFields
+	sizeCache            protoimpl.SizeCache
 }
 
 func (x *PostTranslateModifyRequest) Reset() {
 	*x = PostTranslateModifyRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_service_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_service_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostTranslateModifyRequest) String() string {
@@ -373,7 +353,7 @@ func (*PostTranslateModifyRequest) ProtoMessage() {}
 
 func (x *PostTranslateModifyRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_service_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -412,21 +392,18 @@ func (x *PostTranslateModifyRequest) GetSecrets() []*v33.Secret {
 // PostTranslateModifyResponse is the expected response from an extension and contains
 // the full list of xDS clusters and secrets to be used for the xDS config.
 type PostTranslateModifyResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Clusters      []*v32.Cluster         `protobuf:"bytes,1,rep,name=clusters,proto3" json:"clusters,omitempty"`
+	Secrets       []*v33.Secret          `protobuf:"bytes,2,rep,name=secrets,proto3" json:"secrets,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Clusters []*v32.Cluster `protobuf:"bytes,1,rep,name=clusters,proto3" json:"clusters,omitempty"`
-	Secrets  []*v33.Secret  `protobuf:"bytes,2,rep,name=secrets,proto3" json:"secrets,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *PostTranslateModifyResponse) Reset() {
 	*x = PostTranslateModifyResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_proto_extension_service_proto_msgTypes[7]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_proto_extension_service_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *PostTranslateModifyResponse) String() string {
@@ -437,7 +414,7 @@ func (*PostTranslateModifyResponse) ProtoMessage() {}
 
 func (x *PostTranslateModifyResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_proto_extension_service_proto_msgTypes[7]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -468,7 +445,7 @@ func (x *PostTranslateModifyResponse) GetSecrets() []*v33.Secret {
 
 var File_proto_extension_service_proto protoreflect.FileDescriptor
 
-var file_proto_extension_service_proto_rawDesc = []byte{
+var file_proto_extension_service_proto_rawDesc = string([]byte{
 	0x0a, 0x1d, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x65, 0x78, 0x74, 0x65, 0x6e, 0x73, 0x69, 0x6f,
 	0x6e, 0x2f, 0x73, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12,
 	0x16, 0x65, 0x6e, 0x76, 0x6f, 0x79, 0x67, 0x61, 0x74, 0x65, 0x77, 0x61, 0x79, 0x2e, 0x65, 0x78,
@@ -606,22 +583,22 @@ var file_proto_extension_service_proto_rawDesc = []byte{
 	0x79, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x22, 0x00, 0x42, 0x11, 0x5a, 0x0f, 0x70,
 	0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x65, 0x78, 0x74, 0x65, 0x6e, 0x73, 0x69, 0x6f, 0x6e, 0x62, 0x06,
 	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+})
 
 var (
 	file_proto_extension_service_proto_rawDescOnce sync.Once
-	file_proto_extension_service_proto_rawDescData = file_proto_extension_service_proto_rawDesc
+	file_proto_extension_service_proto_rawDescData []byte
 )
 
 func file_proto_extension_service_proto_rawDescGZIP() []byte {
 	file_proto_extension_service_proto_rawDescOnce.Do(func() {
-		file_proto_extension_service_proto_rawDescData = protoimpl.X.CompressGZIP(file_proto_extension_service_proto_rawDescData)
+		file_proto_extension_service_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_proto_extension_service_proto_rawDesc), len(file_proto_extension_service_proto_rawDesc)))
 	})
 	return file_proto_extension_service_proto_rawDescData
 }
 
 var file_proto_extension_service_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
-var file_proto_extension_service_proto_goTypes = []interface{}{
+var file_proto_extension_service_proto_goTypes = []any{
 	(*PostRouteModifyRequest)(nil),           // 0: envoygateway.extension.PostRouteModifyRequest
 	(*PostRouteModifyResponse)(nil),          // 1: envoygateway.extension.PostRouteModifyResponse
 	(*PostVirtualHostModifyRequest)(nil),     // 2: envoygateway.extension.PostVirtualHostModifyRequest
@@ -676,109 +653,11 @@ func file_proto_extension_service_proto_init() {
 		return
 	}
 	file_proto_extension_context_proto_init()
-	if !protoimpl.UnsafeEnabled {
-		file_proto_extension_service_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostRouteModifyRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_service_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostRouteModifyResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_service_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostVirtualHostModifyRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_service_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostVirtualHostModifyResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_service_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostHTTPListenerModifyRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_service_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostHTTPListenerModifyResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_service_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostTranslateModifyRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_proto_extension_service_proto_msgTypes[7].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*PostTranslateModifyResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_proto_extension_service_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_proto_extension_service_proto_rawDesc), len(file_proto_extension_service_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   8,
 			NumExtensions: 0,
@@ -789,7 +668,6 @@ func file_proto_extension_service_proto_init() {
 		MessageInfos:      file_proto_extension_service_proto_msgTypes,
 	}.Build()
 	File_proto_extension_service_proto = out.File
-	file_proto_extension_service_proto_rawDesc = nil
 	file_proto_extension_service_proto_goTypes = nil
 	file_proto_extension_service_proto_depIdxs = nil
 }
