@@ -12,7 +12,6 @@ import (
 	"path"
 
 	"github.com/spf13/cobra"
-	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	clicfg "sigs.k8s.io/controller-runtime/pkg/client/config"
 
@@ -29,15 +28,15 @@ var overwriteControlPlaneCerts bool
 // TODO: make this path configurable or use server config directly.
 const defaultLocalCertPath = "/tmp/envoy-gateway/certs"
 
-// getCertGenCommand returns the certGen cobra command to be executed.
-func getCertGenCommand() *cobra.Command {
+// GetCertGenCommand returns the certGen cobra command to be executed.
+func GetCertGenCommand() *cobra.Command {
 	var local bool
 
 	cmd := &cobra.Command{
 		Use:   "certgen",
 		Short: "Generate Control Plane Certificates",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return certGen(local)
+			return certGen(cmd.Context(), local)
 		},
 	}
 
@@ -49,7 +48,7 @@ func getCertGenCommand() *cobra.Command {
 }
 
 // certGen generates control plane certificates.
-func certGen(local bool) error {
+func certGen(ctx context.Context, local bool) error {
 	cfg, err := config.New()
 	if err != nil {
 		return err
@@ -68,7 +67,7 @@ func certGen(local bool) error {
 			return fmt.Errorf("failed to create controller-runtime client: %w", err)
 		}
 
-		if err = outputCertsForKubernetes(ctrl.SetupSignalHandler(), cli, cfg, overwriteControlPlaneCerts, certs); err != nil {
+		if err = outputCertsForKubernetes(ctx, cli, cfg, overwriteControlPlaneCerts, certs); err != nil {
 			return fmt.Errorf("failed to output certificates: %w", err)
 		}
 	} else {
