@@ -11,8 +11,8 @@ import (
 	"github.com/envoyproxy/gateway/internal/cmd/version"
 )
 
-// getVersionCommand returns the version cobra command to be executed.
-func getVersionCommand() *cobra.Command {
+// GetVersionCommand returns the version cobra command to be executed.
+func GetVersionCommand() *cobra.Command {
 	var output string
 
 	cmd := &cobra.Command{
