@@ -204,9 +204,10 @@ func (t *Translator) processHTTPRouteRules(httpRoute *HTTPRouteContext, parentRe
 		}
 
 		dstAddrTypeMap := make(map[ir.DestinationAddressType]int)
-
-		for _, backendRef := range rule.BackendRefs {
-			ds, err := t.processDestination(backendRef, parentRef, httpRoute, resources)
+		destName := irRouteDestinationName(httpRoute, ruleIdx)
+		for i, backendRef := range rule.BackendRefs {
+			settingName := irDestinationSettingName(destName, i)
+			ds, err := t.processDestination(settingName, backendRef, parentRef, httpRoute, resources)
 			if !t.IsEnvoyServiceRouting(envoyProxy) && ds != nil && len(ds.Endpoints) > 0 && ds.AddressType != nil {
 				dstAddrTypeMap[*ds.AddressType]++
 			}
@@ -231,7 +232,7 @@ func (t *Translator) processHTTPRouteRules(httpRoute *HTTPRouteContext, parentRe
 
 				if route.Destination == nil {
 					route.Destination = &ir.RouteDestination{
-						Name: irRouteDestinationName(httpRoute, ruleIdx),
+						Name: destName,
 					}
 				}
 				route.Destination.Settings = append(route.Destination.Settings, ds)
@@ -591,8 +592,10 @@ func (t *Translator) processGRPCRouteRules(grpcRoute *GRPCRouteContext, parentRe
 			return nil, err
 		}
 
-		for _, backendRef := range rule.BackendRefs {
-			ds, err := t.processDestination(backendRef, parentRef, grpcRoute, resources)
+		destName := irRouteDestinationName(grpcRoute, ruleIdx)
+		for i, backendRef := range rule.BackendRefs {
+			settingName := irDestinationSettingName(destName, i)
+			ds, err := t.processDestination(settingName, backendRef, parentRef, grpcRoute, resources)
 			if ds == nil {
 				continue
 			}
@@ -613,7 +616,7 @@ func (t *Translator) processGRPCRouteRules(grpcRoute *GRPCRouteContext, parentRe
 
 				if route.Destination == nil {
 					route.Destination = &ir.RouteDestination{
-						Name: irRouteDestinationName(grpcRoute, ruleIdx),
+						Name: destName,
 					}
 				}
 				route.Destination.Settings = append(route.Destination.Settings, ds)
@@ -876,11 +879,13 @@ func (t *Translator) processTLSRouteParentRefs(tlsRoute *TLSRouteContext, resour
 		// not on the Route as a whole.
 		var destSettings []*ir.DestinationSetting
 
+		destName := irRouteDestinationName(tlsRoute, -1 /*rule index*/)
 		// compute backends
 		for _, rule := range tlsRoute.Spec.Rules {
-			for _, backendRef := range rule.BackendRefs {
+			for i, backendRef := range rule.BackendRefs {
+				settingName := irDestinationSettingName(destName, i)
 				// not yet handled, requires to align with the conformance test - TLSRouteInvalidReferenceGrant.
-				ds, _ := t.processDestination(backendRef, parentRef, tlsRoute, resources)
+				ds, _ := t.processDestination(settingName, backendRef, parentRef, tlsRoute, resources)
 				if ds != nil {
 					destSettings = append(destSettings, ds)
 				}
@@ -931,7 +936,7 @@ func (t *Translator) processTLSRouteParentRefs(tlsRoute *TLSRouteContext, resour
 						SNIs: hosts,
 					}},
 					Destination: &ir.RouteDestination{
-						Name:     irRouteDestinationName(tlsRoute, -1 /*rule index*/),
+						Name:     destName,
 						Settings: destSettings,
 					},
 				}
@@ -1019,8 +1024,10 @@ func (t *Translator) processUDPRouteParentRefs(udpRoute *UDPRouteContext, resour
 			continue
 		}
 
-		for _, backendRef := range udpRoute.Spec.Rules[0].BackendRefs {
-			ds, err := t.processDestination(backendRef, parentRef, udpRoute, resources)
+		destName := irRouteDestinationName(udpRoute, -1 /*rule index*/)
+		for i, backendRef := range udpRoute.Spec.Rules[0].BackendRefs {
+			settingName := irDestinationSettingName(destName, i)
+			ds, err := t.processDestination(settingName, backendRef, parentRef, udpRoute, resources)
 			// skip adding the route and provide the reason via route status.
 			if err != nil {
 				routeStatus := GetRouteStatus(udpRoute)
@@ -1034,7 +1041,11 @@ func (t *Translator) processUDPRouteParentRefs(udpRoute *UDPRouteContext, resour
 				)
 				continue
 			}
-			destSettings = append(destSettings, ds)
+
+			// Skip nil destination settings
+			if ds != nil {
+				destSettings = append(destSettings, ds)
+			}
 		}
 
 		// If no negative condition has been set for ResolvedRefs, set "ResolvedRefs=True"
@@ -1074,7 +1085,7 @@ func (t *Translator) processUDPRouteParentRefs(udpRoute *UDPRouteContext, resour
 				irRoute := &ir.UDPRoute{
 					Name: irUDPRouteName(udpRoute),
 					Destination: &ir.RouteDestination{
-						Name:     irRouteDestinationName(udpRoute, -1 /*rule index*/),
+						Name:     destName,
 						Settings: destSettings,
 					},
 				}
@@ -1161,8 +1172,10 @@ func (t *Translator) processTCPRouteParentRefs(tcpRoute *TCPRouteContext, resour
 			continue
 		}
 
-		for _, backendRef := range tcpRoute.Spec.Rules[0].BackendRefs {
-			ds, err := t.processDestination(backendRef, parentRef, tcpRoute, resources)
+		destName := irRouteDestinationName(tcpRoute, -1 /*rule index*/)
+		for i, backendRef := range tcpRoute.Spec.Rules[0].BackendRefs {
+			settingName := irDestinationSettingName(destName, i)
+			ds, err := t.processDestination(settingName, backendRef, parentRef, tcpRoute, resources)
 			// skip adding the route and provide the reason via route status.
 			if err != nil {
 				routeStatus := GetRouteStatus(tcpRoute)
@@ -1176,7 +1189,10 @@ func (t *Translator) processTCPRouteParentRefs(tcpRoute *TCPRouteContext, resour
 				)
 				continue
 			}
-			destSettings = append(destSettings, ds)
+			// Skip nil destination settings
+			if ds != nil {
+				destSettings = append(destSettings, ds)
+			}
 		}
 
 		// If no negative condition has been set for ResolvedRefs, set "ResolvedRefs=True"
@@ -1215,7 +1231,7 @@ func (t *Translator) processTCPRouteParentRefs(tcpRoute *TCPRouteContext, resour
 				irRoute := &ir.TCPRoute{
 					Name: irTCPRouteName(tcpRoute),
 					Destination: &ir.RouteDestination{
-						Name:     irRouteDestinationName(tcpRoute, -1 /*rule index*/),
+						Name:     destName,
 						Settings: destSettings,
 					},
 				}
@@ -1267,7 +1283,7 @@ func (t *Translator) processTCPRouteParentRefs(tcpRoute *TCPRouteContext, resour
 // processDestination translates a backendRef into a destination settings.
 // If an error occurs during this conversion, an error is returned, and the associated routes are expected to become inactive.
 // This will result in a direct 500 response for HTTP-based requests.
-func (t *Translator) processDestination(backendRefContext BackendRefContext,
+func (t *Translator) processDestination(name string, backendRefContext BackendRefContext,
 	parentRef *RouteParentContext, route RouteContext, resources *resource.Resources,
 ) (ds *ir.DestinationSetting, err error) {
 	routeType := GetRouteType(route)
@@ -1320,12 +1336,13 @@ func (t *Translator) processDestination(backendRefContext BackendRefContext,
 		} else {
 			backendIps := resources.GetServiceImport(backendNamespace, string(backendRef.Name)).Spec.IPs
 			for _, ip := range backendIps {
-				ep := ir.NewDestEndpoint(ip, uint32(*backendRef.Port), false)
+				ep := ir.NewDestEndpoint(ip, uint32(*backendRef.Port), false, nil)
 				endpoints = append(endpoints, ep)
 			}
 		}
 
 		ds = &ir.DestinationSetting{
+			Name:        name,
 			Weight:      &weight,
 			Protocol:    protocol,
 			Endpoints:   endpoints,
@@ -1333,7 +1350,7 @@ func (t *Translator) processDestination(backendRefContext BackendRefContext,
 		}
 
 	case resource.KindService:
-		ds = t.processServiceDestinationSetting(backendRef.BackendObjectReference, backendNamespace, protocol, resources, envoyProxy)
+		ds = t.processServiceDestinationSetting(name, backendRef.BackendObjectReference, backendNamespace, protocol, resources, envoyProxy)
 		ds.TLS, err = t.applyBackendTLSSetting(
 			backendRef.BackendObjectReference,
 			backendNamespace,
@@ -1355,9 +1372,12 @@ func (t *Translator) processDestination(backendRefContext BackendRefContext,
 		if err != nil {
 			return nil, err
 		}
-		ds.IPFamily = getServiceIPFamily(resources.GetService(backendNamespace, string(backendRef.Name)))
+		svc := resources.GetService(backendNamespace, string(backendRef.Name))
+		ds.IPFamily = getServiceIPFamily(svc)
+		ds.ZoneAwareRoutingEnabled = isZoneAwareRoutingEnabled(svc)
+
 	case egv1a1.KindBackend:
-		ds = t.processBackendDestinationSetting(backendRef.BackendObjectReference, backendNamespace, protocol, resources)
+		ds = t.processBackendDestinationSetting(name, backendRef.BackendObjectReference, backendNamespace, protocol, resources)
 		ds.TLS, err = t.applyBackendTLSSetting(
 			backendRef.BackendObjectReference,
 			backendNamespace,
@@ -1414,6 +1434,7 @@ func validateDestinationSettings(destinationSettings *ir.DestinationSetting, end
 }
 
 func (t *Translator) processServiceDestinationSetting(
+	name string,
 	backendRef gwapiv1.BackendObjectReference,
 	backendNamespace string,
 	protocol ir.AppProtocol,
@@ -1450,14 +1471,16 @@ func (t *Translator) processServiceDestinationSetting(
 		endpoints, addrType = getIREndpointsFromEndpointSlices(endpointSlices, servicePort.Name, servicePort.Protocol)
 	} else {
 		// Fall back to Service ClusterIP routing
-		ep := ir.NewDestEndpoint(service.Spec.ClusterIP, uint32(*backendRef.Port), false)
+		ep := ir.NewDestEndpoint(service.Spec.ClusterIP, uint32(*backendRef.Port), false, nil)
 		endpoints = append(endpoints, ep)
 	}
 
 	return &ir.DestinationSetting{
-		Protocol:    protocol,
-		Endpoints:   endpoints,
-		AddressType: addrType,
+		Name:                    name,
+		Protocol:                protocol,
+		Endpoints:               endpoints,
+		AddressType:             addrType,
+		ZoneAwareRoutingEnabled: isZoneAwareRoutingEnabled(service),
 	}
 }
 
@@ -1475,6 +1498,22 @@ func getBackendFilters(routeType gwapiv1.Kind, backendRefContext BackendRefConte
 	}
 
 	return nil
+}
+
+func isZoneAwareRoutingEnabled(svc *corev1.Service) bool {
+	if trafficDist := svc.Spec.TrafficDistribution; trafficDist != nil {
+		return *trafficDist == corev1.ServiceTrafficDistributionPreferClose
+	}
+
+	// Allows annotation values that align with Kubernetes defaults.
+	// Ref:
+	// https://kubernetes.io/docs/concepts/services-networking/topology-aware-routing/#enabling-topology-aware-routing
+	// https://github.com/kubernetes/kubernetes/blob/9d9e1afdf78bce0a517cc22557457f942040ca19/staging/src/k8s.io/endpointslice/utils.go#L355-L368
+	if val, ok := svc.Annotations[corev1.AnnotationTopologyMode]; ok {
+		return val == "Auto" || val == "auto"
+	}
+
+	return false
 }
 
 func (t *Translator) processDestinationFilters(routeType gwapiv1.Kind, backendRefContext BackendRefContext, parentRef *RouteParentContext, route RouteContext, resources *resource.Resources) (*ir.DestinationFilters, error) {
@@ -1674,12 +1713,12 @@ func getIREndpointsFromEndpointSlice(endpointSlice *discoveryv1.EndpointSlice, p
 				// Drain the endpoint if it is being terminated
 				draining := *conditions.Terminating
 				for _, address := range endpoint.Addresses {
-					ep := ir.NewDestEndpoint(address, uint32(*endpointPort.Port), draining)
+					ep := ir.NewDestEndpoint(address, uint32(*endpointPort.Port), draining, endpoint.Zone)
 					endpoints = append(endpoints, ep)
 				}
 			} else if conditions.Ready == nil || *conditions.Ready {
 				for _, address := range endpoint.Addresses {
-					ep := ir.NewDestEndpoint(address, uint32(*endpointPort.Port), false)
+					ep := ir.NewDestEndpoint(address, uint32(*endpointPort.Port), false, endpoint.Zone)
 					endpoints = append(endpoints, ep)
 				}
 			}
@@ -1731,7 +1770,7 @@ func getTargetBackendReference(backendRef gwapiv1a2.BackendObjectReference, back
 	return ref
 }
 
-func (t *Translator) processBackendDestinationSetting(backendRef gwapiv1.BackendObjectReference, backendNamespace string, protocol ir.AppProtocol, resources *resource.Resources) *ir.DestinationSetting {
+func (t *Translator) processBackendDestinationSetting(name string, backendRef gwapiv1.BackendObjectReference, backendNamespace string, protocol ir.AppProtocol, resources *resource.Resources) *ir.DestinationSetting {
 	var (
 		dstEndpoints []*ir.DestinationEndpoint
 		dstAddrType  *ir.DestinationAddressType
@@ -1789,6 +1828,7 @@ func (t *Translator) processBackendDestinationSetting(backendRef gwapiv1.Backend
 	}
 
 	ds := &ir.DestinationSetting{
+		Name:        name,
 		Protocol:    protocol,
 		Endpoints:   dstEndpoints,
 		AddressType: dstAddrType,
