@@ -58,7 +58,7 @@ func (i *Infra) CreateOrUpdateProxyInfra(ctx context.Context, infra *ir.Infra) e
 		XdsServerHost:   ptr.To("0.0.0.0"),
 		WasmServerPort:  ptr.To(int32(0)),
 		AdminServerPort: ptr.To(int32(0)),
-		ReadyServerPort: ptr.To(int32(0)),
+		StatsServerPort: ptr.To(int32(0)),
 	}
 
 	args, err := common.BuildProxyArgs(proxyInfra, proxyConfig.Spec.Shutdown, bootstrapConfigOptions, proxyName)
