@@ -6,20 +6,23 @@
 package filters
 
 import (
+	routev3 "github.com/envoyproxy/go-control-plane/envoy/config/route/v3"
 	grpcstats "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/grpc_stats/v3"
 	grpcweb "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/grpc_web/v3"
+	healthcheck "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/health_check/v3"
 	httprouter "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/router/v3"
 	hcm "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
+	matcherv3 "github.com/envoyproxy/go-control-plane/envoy/type/matcher/v3"
 	"github.com/envoyproxy/go-control-plane/pkg/wellknown"
 	"google.golang.org/protobuf/types/known/wrapperspb"
 
-	"github.com/envoyproxy/gateway/internal/utils/protocov"
+	"github.com/envoyproxy/gateway/internal/utils/proto"
 )
 
 var GRPCWeb, GRPCStats *hcm.HttpFilter
 
 func init() {
-	any, err := protocov.ToAnyWithValidation(&grpcweb.GrpcWeb{})
+	any, err := proto.ToAnyWithValidation(&grpcweb.GrpcWeb{})
 	if err != nil {
 		panic(err)
 	}
@@ -30,7 +33,7 @@ func init() {
 		},
 	}
 
-	any, err = protocov.ToAnyWithValidation(&grpcstats.FilterConfig{
+	any, err = proto.ToAnyWithValidation(&grpcstats.FilterConfig{
 		EmitFilterState: true,
 		PerMethodStatSpecifier: &grpcstats.FilterConfig_StatsForAllMethods{
 			StatsForAllMethods: &wrapperspb.BoolValue{Value: true},
@@ -48,7 +51,7 @@ func init() {
 }
 
 func GenerateRouterFilter(enableEnvoyHeaders bool) (*hcm.HttpFilter, error) {
-	any, err := protocov.ToAnyWithValidation(&httprouter.Router{
+	anyCfg, err := proto.ToAnyWithValidation(&httprouter.Router{
 		SuppressEnvoyHeaders: !enableEnvoyHeaders,
 	})
 	if err != nil {
@@ -57,7 +60,34 @@ func GenerateRouterFilter(enableEnvoyHeaders bool) (*hcm.HttpFilter, error) {
 	return &hcm.HttpFilter{
 		Name: wellknown.Router,
 		ConfigType: &hcm.HttpFilter_TypedConfig{
-			TypedConfig: any,
+			TypedConfig: anyCfg,
+		},
+	}, nil
+}
+
+func GenerateHealthCheckFilter(checkPath string) (*hcm.HttpFilter, error) {
+	anyCfg, err := proto.ToAnyWithValidation(&healthcheck.HealthCheck{
+		PassThroughMode: &wrapperspb.BoolValue{Value: false},
+		Headers: []*routev3.HeaderMatcher{
+			{
+				Name: ":path",
+				HeaderMatchSpecifier: &routev3.HeaderMatcher_StringMatch{
+					StringMatch: &matcherv3.StringMatcher{
+						MatchPattern: &matcherv3.StringMatcher_Exact{
+							Exact: checkPath,
+						},
+					},
+				},
+			},
+		},
+	})
+	if err != nil {
+		return nil, err
+	}
+	return &hcm.HttpFilter{
+		Name: wellknown.HealthCheck,
+		ConfigType: &hcm.HttpFilter_TypedConfig{
+			TypedConfig: anyCfg,
 		},
 	}, nil
 }
