@@ -17,7 +17,7 @@ import (
 
 	egv1a1 "github.com/envoyproxy/gateway/api/v1alpha1"
 	"github.com/envoyproxy/gateway/internal/ir"
-	"github.com/envoyproxy/gateway/internal/utils/protocov"
+	"github.com/envoyproxy/gateway/internal/utils/proto"
 	"github.com/envoyproxy/gateway/internal/xds/types"
 )
 
@@ -29,8 +29,12 @@ type basicAuth struct{}
 
 var _ httpFilter = &basicAuth{}
 
-// patchHCM builds and appends the basic_auth Filter to the HTTP Connection Manager
-// if applicable, and it does not already exist.
+// patchHCM updates the HTTPConnectionManager with a Basic Auth HTTP filter for routes requiring authentication.
+// It scans through all routes in the provided HTTPListener, and if a route has a BasicAuth configuration,
+// it checks for the presence of a corresponding filter in the manager. If the filter is not already present,
+// it generates and appends a new Basic Auth filter.
+// The function returns an error if either the HTTPConnectionManager or HTTPListener is nil, or if an error occurs
+// during the filter creation process.
 func (*basicAuth) patchHCM(mgr *hcmv3.HttpConnectionManager, irListener *ir.HTTPListener) error {
 	if mgr == nil {
 		return errors.New("hcm is nil")
@@ -38,34 +42,35 @@ func (*basicAuth) patchHCM(mgr *hcmv3.HttpConnectionManager, irListener *ir.HTTP
 	if irListener == nil {
 		return errors.New("ir listener is nil")
 	}
-	if hcmContainsFilter(mgr, egv1a1.EnvoyFilterBasicAuth.String()) {
-		return nil
-	}
 
-	var (
-		irBasicAuth *ir.BasicAuth
-		filter      *hcmv3.HttpFilter
-		err         error
-	)
+	var errs error
 
 	for _, route := range irListener.Routes {
-		if route.Security != nil && route.Security.BasicAuth != nil {
-			irBasicAuth = route.Security.BasicAuth
-			break
+		if route.Security == nil || route.Security.BasicAuth == nil {
+			continue
 		}
-	}
-	if irBasicAuth == nil {
-		return nil
+
+		// Only generates one Basic Auth Envoy filter for each unique name.
+		// For example, if there are two routes under the same gateway with the
+		// same BasicAuth config, only one BasicAuth filter will be generated.
+		if hcmContainsFilter(mgr, basicAuthFilterName(route.Security.BasicAuth)) {
+			continue
+		}
+
+		filter, err := buildHCMBasicAuthFilter(route.Security.BasicAuth)
+		if err != nil {
+			errs = errors.Join(errs, err)
+			continue
+		}
+
+		mgr.HttpFilters = append(mgr.HttpFilters, filter)
 	}
 
-	// We use the first route that contains the basicAuth config to build the filter.
-	// The HCM-level filter config doesn't matter since it is overridden at the route level.
-	if filter, err = buildHCMBasicAuthFilter(irBasicAuth); err != nil {
-		return err
-	}
-	mgr.HttpFilters = append(mgr.HttpFilters, filter)
-	return err
+	return errs
 }
+
+// patchHCM builds and appends the basic_auth Filter to the HTTP Connection Manager
+// if applicable, and it does not already exist.
 
 // buildHCMBasicAuthFilter returns a basic_auth HTTP filter from the provided IR HTTPRoute.
 func buildHCMBasicAuthFilter(basicAuth *ir.BasicAuth) (*hcmv3.HttpFilter, error) {
@@ -82,20 +87,26 @@ func buildHCMBasicAuthFilter(basicAuth *ir.BasicAuth) (*hcmv3.HttpFilter, error)
 			},
 		},
 	}
-	if err = basicAuthProto.ValidateAll(); err != nil {
-		return nil, err
+	// Set the ForwardUsernameHeader field if it is specified.
+	if basicAuth.ForwardUsernameHeader != nil && *basicAuth.ForwardUsernameHeader != "" {
+		basicAuthProto.ForwardUsernameHeader = *basicAuth.ForwardUsernameHeader
 	}
-	if basicAuthAny, err = protocov.ToAnyWithValidation(basicAuthProto); err != nil {
+
+	if basicAuthAny, err = proto.ToAnyWithValidation(basicAuthProto); err != nil {
 		return nil, err
 	}
 
 	return &hcmv3.HttpFilter{
-		Name: egv1a1.EnvoyFilterBasicAuth.String(),
+		Name: basicAuthFilterName(basicAuth),
 		ConfigType: &hcmv3.HttpFilter_TypedConfig{
 			TypedConfig: basicAuthAny,
 		},
 		Disabled: true,
 	}, nil
+}
+
+func basicAuthFilterName(basicAuth *ir.BasicAuth) string {
+	return perRouteFilterName(egv1a1.EnvoyFilterBasicAuth, basicAuth.Name)
 }
 
 func (*basicAuth) patchResources(*types.ResourceVersionTable, []*ir.HTTPRoute) error {
@@ -120,9 +131,9 @@ func (*basicAuth) patchRoute(route *routev3.Route, irRoute *ir.HTTPRoute) error 
 		basicAuthAny *anypb.Any
 		err          error
 	)
-
+	filterName := basicAuthFilterName(irRoute.Security.BasicAuth)
 	perFilterCfg = route.GetTypedPerFilterConfig()
-	if _, ok := perFilterCfg[egv1a1.EnvoyFilterBasicAuth.String()]; ok {
+	if _, ok := perFilterCfg[filterName]; ok {
 		// This should not happen since this is the only place where the filter
 		// config is added in a route.
 		return fmt.Errorf("route already contains filter config: %s, %+v",
@@ -131,18 +142,14 @@ func (*basicAuth) patchRoute(route *routev3.Route, irRoute *ir.HTTPRoute) error 
 
 	// Overwrite the HCM level filter config with the per route filter config.
 	basicAuthProto := basicAuthPerRouteConfig(irRoute.Security.BasicAuth)
-	if err = basicAuthProto.ValidateAll(); err != nil {
-		return err
-	}
-
-	if basicAuthAny, err = protocov.ToAnyWithValidation(basicAuthProto); err != nil {
+	if basicAuthAny, err = proto.ToAnyWithValidation(basicAuthProto); err != nil {
 		return err
 	}
 
 	if perFilterCfg == nil {
 		route.TypedPerFilterConfig = make(map[string]*anypb.Any)
 	}
-	route.TypedPerFilterConfig[egv1a1.EnvoyFilterBasicAuth.String()] = basicAuthAny
+	route.TypedPerFilterConfig[filterName] = basicAuthAny
 
 	return nil
 }
