@@ -25,7 +25,7 @@ import (
 
 	egv1a1 "github.com/envoyproxy/gateway/api/v1alpha1"
 	"github.com/envoyproxy/gateway/internal/ir"
-	"github.com/envoyproxy/gateway/internal/utils/protocov"
+	"github.com/envoyproxy/gateway/internal/utils/proto"
 	"github.com/envoyproxy/gateway/internal/xds/types"
 )
 
@@ -78,16 +78,11 @@ func (c *customResponse) patchHCM(mgr *hcmv3.HttpConnectionManager, irListener *
 
 // buildHCMCustomResponseFilter returns an OAuth2 HTTP filter from the provided IR HTTPRoute.
 func (c *customResponse) buildHCMCustomResponseFilter(ro *ir.ResponseOverride) (*hcmv3.HttpFilter, error) {
-	proto, err := c.customResponseConfig(ro)
+	config, err := c.customResponseConfig(ro)
 	if err != nil {
 		return nil, err
 	}
-
-	if err := proto.ValidateAll(); err != nil {
-		return nil, err
-	}
-
-	any, err := protocov.ToAnyWithValidation(proto)
+	any, err := proto.ToAnyWithValidation(config)
 	if err != nil {
 		return nil, err
 	}
@@ -239,7 +234,7 @@ func (c *customResponse) buildHTTPAttributeCELInput() (*cncfv3.TypedExtensionCon
 		err error
 	)
 
-	if pb, err = protocov.ToAnyWithValidation(&matcherv3.HttpAttributesCelMatchInput{}); err != nil {
+	if pb, err = proto.ToAnyWithValidation(&matcherv3.HttpAttributesCelMatchInput{}); err != nil {
 		return nil, err
 	}
 
@@ -255,7 +250,7 @@ func (c *customResponse) buildStatusCodeInput() (*cncfv3.TypedExtensionConfig, e
 		err error
 	)
 
-	if pb, err = protocov.ToAnyWithValidation(&envoymatcherv3.HttpResponseStatusCodeMatchInput{}); err != nil {
+	if pb, err = proto.ToAnyWithValidation(&envoymatcherv3.HttpResponseStatusCodeMatchInput{}); err != nil {
 		return nil, err
 	}
 
@@ -362,11 +357,7 @@ func (c *customResponse) buildStatusCodeCELMatcher(codeRange ir.StatusCodeRange)
 			},
 		},
 	}
-	if err := matcher.ValidateAll(); err != nil {
-		return nil, err
-	}
-
-	if pb, err = protocov.ToAnyWithValidation(matcher); err != nil {
+	if pb, err = proto.ToAnyWithValidation(matcher); err != nil {
 		return nil, err
 	}
 
@@ -405,11 +396,7 @@ func (c *customResponse) buildAction(r ir.ResponseOverrideRule) (*matcherv3.Matc
 		err error
 	)
 
-	if err := response.ValidateAll(); err != nil {
-		return nil, err
-	}
-
-	if pb, err = protocov.ToAnyWithValidation(response); err != nil {
+	if pb, err = proto.ToAnyWithValidation(response); err != nil {
 		return nil, err
 	}
 
