@@ -21,7 +21,7 @@ import (
 
 	egv1a1 "github.com/envoyproxy/gateway/api/v1alpha1"
 	"github.com/envoyproxy/gateway/internal/ir"
-	"github.com/envoyproxy/gateway/internal/utils/protocov"
+	"github.com/envoyproxy/gateway/internal/utils/proto"
 	"github.com/envoyproxy/gateway/internal/xds/types"
 )
 
@@ -79,11 +79,7 @@ func buildHCMOAuth2Filter(oidc *ir.OIDC) (*hcmv3.HttpFilter, error) {
 		return nil, err
 	}
 
-	if err := oauth2Proto.ValidateAll(); err != nil {
-		return nil, err
-	}
-
-	OAuth2Any, err := protocov.ToAnyWithValidation(oauth2Proto)
+	OAuth2Any, err := proto.ToAnyWithValidation(oauth2Proto)
 	if err != nil {
 		return nil, err
 	}
@@ -352,8 +348,9 @@ func createOAuth2TokenEndpointCluster(tCtx *types.ResourceVersionTable,
 	ds = &ir.DestinationSetting{
 		Weight: ptr.To[uint32](1),
 		Endpoints: []*ir.DestinationEndpoint{
-			ir.NewDestEndpoint(cluster.hostname, cluster.port, false),
+			ir.NewDestEndpoint(cluster.hostname, cluster.port, false, nil),
 		},
+		Name: destinationSettingName(cluster.name),
 	}
 
 	clusterArgs := &xdsClusterArgs{
