@@ -85,6 +85,12 @@ func clusterName(host string, port uint32) string {
 	return fmt.Sprintf("%s_%d", strings.ReplaceAll(host, ".", "_"), port)
 }
 
+func destinationSettingName(destName string) string {
+	// -1 is used here since this function is used to generate a name
+	// for a backend that is defined using a scalar field that has no index.
+	return fmt.Sprintf("%s/backend/-1", destName)
+}
+
 // enableFilterOnRoute enables a filterType on the provided route.
 func enableFilterOnRoute(route *routev3.Route, filterName string) error {
 	if route == nil {
@@ -180,7 +186,8 @@ func addClusterFromURL(url string, tCtx *types.ResourceVersionTable) error {
 
 	ds = &ir.DestinationSetting{
 		Weight:    ptr.To[uint32](1),
-		Endpoints: []*ir.DestinationEndpoint{ir.NewDestEndpoint(uc.hostname, uc.port, false)},
+		Endpoints: []*ir.DestinationEndpoint{ir.NewDestEndpoint(uc.hostname, uc.port, false, nil)},
+		Name:      destinationSettingName(uc.name),
 	}
 
 	clusterArgs := &xdsClusterArgs{
