@@ -5,7 +5,7 @@
 
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
+// - protoc-gen-go-grpc v1.5.1
 // - protoc             (unknown)
 // source: proto/extension/service.proto
 
@@ -20,8 +20,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	EnvoyGatewayExtension_PostRouteModify_FullMethodName        = "/envoygateway.extension.EnvoyGatewayExtension/PostRouteModify"
@@ -70,8 +70,9 @@ func NewEnvoyGatewayExtensionClient(cc grpc.ClientConnInterface) EnvoyGatewayExt
 }
 
 func (c *envoyGatewayExtensionClient) PostRouteModify(ctx context.Context, in *PostRouteModifyRequest, opts ...grpc.CallOption) (*PostRouteModifyResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostRouteModifyResponse)
-	err := c.cc.Invoke(ctx, EnvoyGatewayExtension_PostRouteModify_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EnvoyGatewayExtension_PostRouteModify_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -79,8 +80,9 @@ func (c *envoyGatewayExtensionClient) PostRouteModify(ctx context.Context, in *P
 }
 
 func (c *envoyGatewayExtensionClient) PostVirtualHostModify(ctx context.Context, in *PostVirtualHostModifyRequest, opts ...grpc.CallOption) (*PostVirtualHostModifyResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostVirtualHostModifyResponse)
-	err := c.cc.Invoke(ctx, EnvoyGatewayExtension_PostVirtualHostModify_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EnvoyGatewayExtension_PostVirtualHostModify_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -88,8 +90,9 @@ func (c *envoyGatewayExtensionClient) PostVirtualHostModify(ctx context.Context,
 }
 
 func (c *envoyGatewayExtensionClient) PostHTTPListenerModify(ctx context.Context, in *PostHTTPListenerModifyRequest, opts ...grpc.CallOption) (*PostHTTPListenerModifyResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostHTTPListenerModifyResponse)
-	err := c.cc.Invoke(ctx, EnvoyGatewayExtension_PostHTTPListenerModify_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EnvoyGatewayExtension_PostHTTPListenerModify_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -97,8 +100,9 @@ func (c *envoyGatewayExtensionClient) PostHTTPListenerModify(ctx context.Context
 }
 
 func (c *envoyGatewayExtensionClient) PostTranslateModify(ctx context.Context, in *PostTranslateModifyRequest, opts ...grpc.CallOption) (*PostTranslateModifyResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(PostTranslateModifyResponse)
-	err := c.cc.Invoke(ctx, EnvoyGatewayExtension_PostTranslateModify_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, EnvoyGatewayExtension_PostTranslateModify_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -107,7 +111,7 @@ func (c *envoyGatewayExtensionClient) PostTranslateModify(ctx context.Context, i
 
 // EnvoyGatewayExtensionServer is the server API for EnvoyGatewayExtension service.
 // All implementations must embed UnimplementedEnvoyGatewayExtensionServer
-// for forward compatibility
+// for forward compatibility.
 type EnvoyGatewayExtensionServer interface {
 	// PostRouteModify provides a way for extensions to modify a route generated by Envoy Gateway before it is finalized.
 	// Doing so allows extensions to configure/modify route fields configured by Envoy Gateway and also to configure the
@@ -137,9 +141,12 @@ type EnvoyGatewayExtensionServer interface {
 	mustEmbedUnimplementedEnvoyGatewayExtensionServer()
 }
 
-// UnimplementedEnvoyGatewayExtensionServer must be embedded to have forward compatible implementations.
-type UnimplementedEnvoyGatewayExtensionServer struct {
-}
+// UnimplementedEnvoyGatewayExtensionServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedEnvoyGatewayExtensionServer struct{}
 
 func (UnimplementedEnvoyGatewayExtensionServer) PostRouteModify(context.Context, *PostRouteModifyRequest) (*PostRouteModifyResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method PostRouteModify not implemented")
@@ -154,6 +161,7 @@ func (UnimplementedEnvoyGatewayExtensionServer) PostTranslateModify(context.Cont
 	return nil, status.Errorf(codes.Unimplemented, "method PostTranslateModify not implemented")
 }
 func (UnimplementedEnvoyGatewayExtensionServer) mustEmbedUnimplementedEnvoyGatewayExtensionServer() {}
+func (UnimplementedEnvoyGatewayExtensionServer) testEmbeddedByValue()                               {}
 
 // UnsafeEnvoyGatewayExtensionServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to EnvoyGatewayExtensionServer will
@@ -163,6 +171,13 @@ type UnsafeEnvoyGatewayExtensionServer interface {
 }
 
 func RegisterEnvoyGatewayExtensionServer(s grpc.ServiceRegistrar, srv EnvoyGatewayExtensionServer) {
+	// If the following call pancis, it indicates UnimplementedEnvoyGatewayExtensionServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&EnvoyGatewayExtension_ServiceDesc, srv)
 }
 
